moved testing into respective files
