+++ internal/overflow/offset_test.go
@@ -1,0 +1,55 @@
+package overflow
+
+import (
+    "testing"
+)
+
+// ...
+func TestFindSubPatternEmpty(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternEmptyWithLength(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternSmallValid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternSmallWithLengthValid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternSmallInvalid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternSmallWithLengthInvalid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternLargeValid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternLargeWithLengthValid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternLargeInvalid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
+
+// ...
+func TestFindSubPatternLargeWithLengthInvalid(t *testing.T) {
+    t.Errorf("test not implemented")
+}
